RavenDB-11310 - Move Elector.ShouldGrantVote to ElectorCastVoteInTermWithShouldGrantVoteCommand
