+++ src/Raven.Server/Documents/Handlers/BatchRequestParser.cs
@@ -10,7 +10,6 @@ using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Raven.Client.Documents.Attachments;
 using Raven.Client.Documents.Commands.Batches;
-using Raven.Client.Documents.Operations;
 using Raven.Client.Documents.Operations.Attachments;
 using Raven.Client.Documents.Operations.Counters;
 using Raven.Client.Documents.Operations.TimeSeries;
@@ -19,7 +18,7 @@ using Raven.Client.Documents.Smuggler;
 using Raven.Client.Util;
 using Raven.Server.Documents.Patch;
 using Raven.Server.Documents.TransactionMerger.Commands;
-using Raven.Server.Documents.PeriodicBackup;using Raven.Server.Exceptions;
+using Raven.Server.Exceptions;
 using Raven.Server.ServerWide;
 using Sparrow;
 using Sparrow.Json;
